Extract logic from factory to functions
