+++ packages/docs/pages/Unfolding.mdx
@@ -7,7 +7,7 @@ Furthermore it will pass on any node params which will act as *global* params in
 
 ```ts
 // When creating the execution
-const { diagram } = ExecutableDiagramFactory.create(original, definitions)
+const { diagram } = createExecutableDiagram(original, definitions)
 ```
 Let's consider an example!
